user-story-tests/29-markdown-component-with-author-defined-lower-priority-shadow-layer.html

<!DOCTYPE html>
<html lang="en">

<head>
    <meta charset="UTF-8">
    <meta name="viewport" content="width=device-width, initial-scale=1.0">
    <title>Markdown component with author-defined lower priority shadow layer</title>
    <script src='./resources/testharness.js'></script>
    <script src='./resources/testharnessreport.js'></script>
</head>

<body>

    <script type="module">
        import { MDBlockWithAuthorDefinedLowerPriorityShadowLayer } from "./components/MDBlockWithAuthorDefinedLowerPriorityShadowLayer.js";
        customElements.define(
            "md-block-with-author-defined-lower-priority-shadow-layer",
            MDBlockWithAuthorDefinedLowerPriorityShadowLayer
        );
    </script>

    <style>
        h2 {
            border: thick dashed red;
        }
    </style>
    <h2>H2 outside shadow tree</h2>
    <md-block-with-author-defined-lower-priority-shadow-layer>
        <script type="text/markdown">
## H2 from markdown (thick dashed red) (from page styles)

### H3 from markdown (thick solid black) (from web component default styles)
    </script>
    </md-block-with-author-defined-lower-priority-shadow-layer>


    <script type="module">
        import ShadowLayers from '../ShadowLayers.js';

        window.onload = function () {

            test(function () {

                assert_true(true, "always true, just to show the test is running");

            }, "Markdown component with author-defined lower priority shadow layer");

        }
    </script>

    <div id="log"></div>

</body>

</html>
